6-size.c

+++ hello_world/6-size.c
@@ -8,10 +8,10 @@
 
 int main(void)
 {
-printf("Size of char: %ld bytes\n", sizeof(char));
-printf("Size of int: %ld bytes\n", sizeof(int));
-printf("Size of long int: %ld bytes\n", sizeof(long int));
-printf("Size of long long int : %ld bytes\n", sizeof(unsigned long int));
-printf("Size of float: %ld bytes\n", sizeof(float));
+printf("Size of char: %d bytes\n", sizeof(char));
+printf("Size of int: %d bytes\n", sizeof(int));
+printf("Size of long int: %d bytes\n", sizeof(long int));
+printf("Size of long long int : %d bytes\n", sizeof(long long int));
+printf("Size of float: %d bytes\n", sizeof(float));
 return (0);
 }
